pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "naimed-kepler-bot"
version = "0.1.0"
description = "Internal Discord bot for employee rewards and performance checks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.5.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

